refactored and other stuff to really get the wow factor
Fixed whitespace and removed unused function
